move old models tests to tools
