+++ gpkit/tests/run_tests.py
@@ -26,9 +26,6 @@ TESTS += t_geometric_program.TESTS
 from gpkit.tests import t_gp_solution_array
 TESTS += t_gp_solution_array.TESTS
 
-from gpkit.tests import t_models
-TESTS += t_models.TESTS
-
 
 import gpkit
